migrated persistence functionality into navbar component
